backend/requirements.txt
fastapi
uvicorn
docker
mysql-connector-python